+++ tests/unit/api_client/test_user_folder_api.py
@@ -9,18 +9,18 @@ class TestUserFolderApi(ApiTestBase):
         return UserFolderApi(session, host)
 
     def test_get_folder__already_set(self, api):
-        api._user_folder = "folder"
-        assert "folder" == api.get_folder()
+        api._user = "me"
+        assert "/Users/me/dbt_python_models/catalog/schema/" == api.get_folder("catalog", "schema")
 
     def test_get_folder__non_200(self, api, session):
-        self.assert_non_200_raises_error(lambda: api.get_folder(), session)
+        self.assert_non_200_raises_error(lambda: api.get_folder("catalog", "schema"), session)
 
     def test_get_folder__200(self, api, session, host):
         session.get.return_value.status_code = 200
         session.get.return_value.json.return_value = {"userName": "me@gmail.com"}
-        folder = api.get_folder()
-        assert folder == "/Users/me@gmail.com"
-        assert api._user_folder == folder
+        folder = api.get_folder("catalog", "schema")
+        assert folder == "/Users/me@gmail.com/dbt_python_models/catalog/schema/"
+        assert api._user == "me@gmail.com"
         session.get.assert_called_once_with(
             f"https://{host}/api/2.0/preview/scim/v2/Me", json=None, params=None
         )