+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## dbt-databricks 1.9.0 (TBD)
 
 - Add support for serverless job clusters on python models ([706](https://github.com/databricks/dbt-databricks/pull/706))
+- Add 'user_folder_for_python' config to switch writing python model notebooks to the user's folder ([706](https://github.com/databricks/dbt-databricks/pull/706))
 
 ## dbt-databricks 1.8.2 (June 24, 2024)